requirements.txt
Flask==2.0.1
PyGithub==1.55
python-dotenv==0.19.0
gunicorn==20.1.0
Flask-WTF==0.15.1
requests==2.26.0
python-dateutil==2.8.2